dynamic loading example added

+++ dynamic_loading/app/comp.sh
@@ -7,12 +7,23 @@ done
 
 for name in app app_static app_dyn; do
     echo "Deleting ${name}"
-    rm "${name}"
+    [[ -f "${name}" ]] && rm "${name}"
 done
 
-g++ -Wall -c main.cpp -I/home/thiago/src/teste/lib
-g++ -o app main.o -L/home/thiago/src/teste/lib -lstateful_writer
+# Compiles the source code and generates the object file main.o
+# ../lib points to the location of the file stateful_writer.h
+g++ -Wall -c main.cpp -I../lib
+
+# Generates the executable binary app_static including the static library libstateful_writer.a
+# Adding this .a file is the same as adding all .o files contained in this library
 g++ -o app_static main.o ../lib/libstateful_writer.a
 
-g++ -I/home/thiago/src/teste/lib -Wall -c main_dyn.cpp
-g++ -rdynamic -Wall main_dyn.o -ldl -o app_dyn
+# Generates a executable binary app using the shared library libstateful_writer.so
+g++ -o app main.o -L../lib -lstateful_writer
+
+# Compiles the source code and generates the object file main_dyn.o
+# ../lib points to the location of the file stateful_writer_imports.h
+g++ -I../lib -Wall -c main_dyn.cpp
+
+# Generates a executable binary app without linking any library
+g++ -Wall main_dyn.o -ldl -o app_dyn
